requirements.txt
dash==1.20.0
dash-renderer
dash-html-components
dash-core-components
dash-table-experiments
dash-cytoscape==0.2.0
dash_extensions
plotly==4.14.3
regex
numpy
pandas
flask==1.1.2
flask-caching
requests
pymongo
scipy
openpyxl
wordcloud
opencv-python